fixing signals

+++ signals.c
@@ -6,7 +6,7 @@
 /*   By: Mathis <Mathis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/02 11:48:52 by Mathis            #+#    #+#             */
-/*   Updated: 2020/10/08 16:30:27 by Mathis           ###   ########.fr       */
+/*   Updated: 2020/10/12 13:36:18 by Mathis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ extern t_printf g_printf;
 
 void    signal_quit(int signal)
 {
-    if(g_shell.pid)
+    if(g_shell.pid && !g_shell.d)
     {
         kill(g_shell.pid, signal);
         g_shell.status = 131;
